+++ packages/ui/CHANGELOG.md
@@ -6,7 +6,7 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 ### Added
 
-- `ui/IconButton` prop `flip` to flip the icon horizontally or vertically
+- `ui/Icon` and `ui/IconButton` prop `flip` to flip the icon horizontally, vertically or both
 
 ## [3.0.0-beta.2] - 2019-07-23